Reconstructor

+++ src/task/import/asterix/asteriximporttask.h
@@ -74,7 +74,7 @@ public:
 
     unsigned int num_packets_overload_ {60};
 
-    float override_tod_offset_{0};
+    double override_tod_offset_{0};
 
     float filter_tod_min_{0};
     float filter_tod_max_{0};
